requirements.txt
aiogram==3.20.0
python_dotenv==1.1.0
aiohttp==3.11.18
